+++ CHANGES.md
@@ -1,8 +1,12 @@
+## Changes in 0.4.0 (in development)
+
+### Enhancements 
+* Added ability to write xcube datasets in zarr format into s3 bucket using
+  xcube.core.dsio.write_cube() #224
+
 ## Changes in 0.3.0
 
 ### New
-* Added ability to write xcube datasets in zarr format into s3 bucket using
-  xcube.core.dsio.write_cube() #224
 * Added new parameter in `xcube gen` called `--no_sort`. Using `--no_sort`, 
   the input file list wont be sorted before creating the xcube dataset. 
   If `--no_sort` parameter is passed, order the input list will be kept.